refactor: move file
Move `Registry` type to own file, since it is used by many modules.
