+++ src/cli/cmd-search.ts
@@ -2,7 +2,6 @@ import log from "./logger";
 import * as os from "os";
 import { EnvParseError, parseEnv } from "../utils/env";
 import { CmdOptions } from "./options";
-import { Registry } from "../services/add-user-service";
 import { formatAsTable } from "./output-formatting";
 import { AsyncResult, Ok, Result } from "ts-results-es";
 import { HttpErrorBase } from "npm-registry-fetch";
@@ -11,6 +10,7 @@ import {
   SearchedPackument,
   SearchService,
 } from "../services/search-service";
+import { Registry } from "../domain/registry";
 
 export type SearchError = EnvParseError | HttpErrorBase;
