+++ src/dependency-resolving.ts
@@ -14,9 +14,9 @@ import {
 import { unityRegistryUrl } from "./domain/registry-url";
 import { recordEntries } from "./utils/record-utils";
 import assert from "assert";
-import { Registry } from "./services/add-user-service";
 import { PackumentNotFoundError } from "./common-errors";
 import { PackumentFetchService } from "./services/fetch-packument";
+import { Registry } from "./domain/registry";
 
 export type DependencyBase = {
   /**
