+++ src/utils/env.ts
@@ -11,7 +11,6 @@ import fs from "fs";
 import { coerceRegistryUrl, makeRegistryUrl } from "../domain/registry-url";
 import { tryGetAuthForRegistry, UPMConfig } from "../domain/upm-config";
 import { CmdOptions } from "../cli/options";
-import { Registry } from "../services/add-user-service";
 import { FileParseError } from "../common-errors";
 import { Err, Ok, Result } from "ts-results-es";
 import {
@@ -25,6 +24,7 @@ import {
   ReleaseVersion,
   tryParseEditorVersion,
 } from "../domain/editor-version";
+import { Registry } from "../domain/registry";
 
 export type Env = Readonly<{
   cwd: string;
