+++ test/wsl.test.ts
@@ -1,7 +1,7 @@
 import * as processModule from "../src/utils/process";
+import { ChildProcessError } from "../src/utils/process";
 import { tryGetWslPath } from "../src/io/wls";
 import { Err, Ok } from "ts-results-es";
-import { ChildProcessError } from "../src/utils/process";
 
 jest.mock("is-wsl", () => ({
   __esModule: true,
